Rename Cognito module to Accounts
The cognito integration replaces the old django accounts approach for user account authentication and authorization. However naming the module cognito is limiting in its scope. The module should be reworked to allow alternative federation services or even a local database for easier testing or smaller deployment. This commit will start us down that path.
