+++ src/accounts/apps.py
@@ -1,0 +1,4 @@
+from django.apps import AppConfig
+
+class AccountsConfig(AppConfig):
+    name = 'accounts'
