--- src/cognito/apps.py
@@ -1,4 +1,0 @@
-from django.apps import AppConfig
-
-class CognitoConfig(AppConfig):
-    name = 'cognito'
