+++ src/dataloaderinterface/models.py
@@ -16,7 +16,7 @@ from dataloaderinterface.querysets import SiteRegistrationQuerySet, SensorOutput
 
 
 #TODO: replace with a different model approach as this has been deprecated
-import cognito.models
+import accounts.models
 
 class SiteRegistration(models.Model):
     registration_id = models.AutoField(primary_key=True, db_column='RegistrationID')
@@ -25,7 +25,7 @@ class SiteRegistration(models.Model):
     registration_date = models.DateTimeField(db_column='RegistrationDate', default=datetime.utcnow)
     deployment_date = models.DateTimeField(db_column='DeploymentDate', blank=True, null=True)
 
-    account_id = models.ForeignKey(cognito.models.Account, on_delete=models.CASCADE, db_column='account_id', related_name='deployed_sites')
+    account_id = models.ForeignKey(accounts.models.Account, on_delete=models.CASCADE, db_column='account_id', related_name='deployed_sites')
     affiliation_id = models.IntegerField(db_column='AffiliationID')
 
     person_id = models.IntegerField(db_column='PersonID', null=True)
@@ -54,8 +54,8 @@ class SiteRegistration(models.Model):
     site_notes = models.TextField(db_column='SiteNotes', blank=True, null=True)
 
     #TODO: VERIFY THIS BEHAVIOR
-    followed_by = models.ManyToManyField(cognito.models.Account, related_name='followed_sites')
-    alert_listeners = models.ManyToManyField(cognito.models.Account, related_name='+', through='SiteAlert')
+    followed_by = models.ManyToManyField(accounts.models.Account, related_name='followed_sites')
+    alert_listeners = models.ManyToManyField(accounts.models.Account, related_name='+', through='SiteAlert')
 
     objects = SiteRegistrationQuerySet.as_manager()
 
@@ -197,7 +197,7 @@ class SiteSensor(models.Model):
 
 
 class SiteAlert(models.Model):
-    account_id = models.ForeignKey(cognito.models.Account ,db_column='account_id', on_delete=models.CASCADE, related_name='site_alerts')
+    account_id = models.ForeignKey(accounts.models.Account ,db_column='account_id', on_delete=models.CASCADE, related_name='site_alerts')
     site_registration = models.ForeignKey('SiteRegistration', db_column='RegistrationID', on_delete=models.CASCADE, related_name='alerts')
     last_alerted = models.DateTimeField(db_column='LastAlerted', blank=True, null=True)
     hours_threshold = models.DurationField(db_column='HoursThreshold', default=timedelta(hours=1))
