+++ src/leafpack/models.py
@@ -8,7 +8,7 @@ from operator import __or__ as OR
 
 from functools import reduce
 
-import cognito
+import accounts
 
 class Macroinvertebrate(models.Model):
     """
@@ -250,7 +250,7 @@ class LeafPackType(models.Model):
 
     name = models.CharField(max_length=255, unique=True)
     #TODO: remove auth dependency on django models
-    created_by = models.ForeignKey(cognito.models.Account, on_delete=models.CASCADE, blank=True, null=True)
+    created_by = models.ForeignKey(accounts.models.Account, on_delete=models.CASCADE, blank=True, null=True)
 
     def __str__(self):
         return self.name
